refactor: create Header component for /webtime
