+++ src/components/NewTab/index.tsx
@@ -18,7 +18,6 @@ import {
 import { db } from "../../../firebase-config";
 import { sortTabs } from "../../utils/firestore";
 import ToggleViewBtn from "./ToggleViewBtn";
-import { useDateStore } from "../../store";
 
 export interface Tab extends chrome.tabs.Tab {
   lastAccessed: FieldValue;
@@ -54,10 +53,6 @@ const NewTab = () => {
   const [tabOrder, setTabOrder] = useState<number[]>([]);
   const [isTabsGrid, setIsTabsGrid] = useState<boolean>(false);
   const [currentWindowId, setCurrentWindowId] = useState<number>(0);
-  const date = useDateStore((state) => state.date);
-  const increaseDate = useDateStore((state) => state.increaseDate);
-  const decreaseDate = useDateStore((state) => state.decreaseDate);
-  // console.log("current order", tabOrder);
   console.log("current windowId", currentWindowId);
   console.log("current tabs", tabs);
   console.log("current spaces", spaces);
@@ -66,10 +61,6 @@ const NewTab = () => {
   );
   const location = useLocation();
   const newSpaceInputRef = useRef<HTMLInputElement>(null);
-  // console.log(
-  //   "spaces",
-  //   spaces.map((space) => space.id),
-  // );
   const tabOrderRef = useRef<number[]>(tabOrder);
   tabOrderRef.current = tabOrder;
   useEffect(() => {
@@ -260,7 +251,7 @@ const NewTab = () => {
       chrome.runtime.onMessage.removeListener(handleMessagePassing);
     };
   }, [currentWindowId]);
- 
+
   function openLink(
     e: React.MouseEvent<HTMLAnchorElement, MouseEvent>,
     tab: Tab,
@@ -544,24 +535,6 @@ const NewTab = () => {
                   <CopyToClipboard onCopySpaceLink={copySpaceLink} />
                 )}
               </>
-            )}
-            {location.pathname === "/webtime" && (
-              <div className="flex w-full items-center">
-                <h1 className="text-3xl font-bold">
-                  Your Time Spent on Websites
-                </h1>
-                <div className="ml-auto flex items-center justify-center gap-2">
-                  <button
-                    className="h-8 w-8 text-4xl"
-                    onClick={decreaseDate}
-                  >{`<`}</button>
-                  <p className="grow text-3xl">{date}</p>
-                  <button
-                    className="h-8 w-8 text-4xl"
-                    onClick={increaseDate}
-                  >{`>`}</button>
-                </div>
-              </div>
             )}
           </div>
           <Outlet />
